requirements.txt
Pillow
openai
pytesseract